remove explore page
The explore page (at `/explore`) showed recommended extensions based on a hardcoded list, selected Sourcegraph editor/browser/CLI integrations, and a few repositories on the instance. It was intended to be a good place for users to visit to get set up with Sourcegraph. In practice, we never promoted it to users, never included it in our demo paths, and never spent any time maintaining it. We also never heard from users that it was helpful (or not). Based on this (intuition suggests it's not useful + lack of data to suggest otherwise), we should remove it.

Its intent is good and we should find a good way to accomplish its goals in the product, but good intent is not enough for a product feature to stick around (especially when it occupies a prominent link in the global nav).

- fix #5318
- fix #4019